Rename servlets -> http
